Restore movement and weapons function
A good halfway to getting back to where I was with this new character base. Thankfully it didn't take to long to glue everything back together, and I am feeling the silly ambiguous-ness of adding a method that calls another method and having a gel layer of commands with very simple data, but I see the appeal the more I work on it.